Time: 226 ms (14.96%), Space: 37.5 MB (30.63%) - LeetHub

+++ 1383-maximum-performance-of-a-team/1383-maximum-performance-of-a-team.cpp
@@ -3,7 +3,6 @@ using min_pq = priority_queue<T, vector<T>, greater<T>>;
 using ll = long long;
 using pii = pair<ll,ll>;
 using vp = vector<pii>;
-
 const int mod = 1e9+7;
 void chmin(ll &x, ll y) {x = min(x,y);}
 void chmax(ll &x, ll y) {x = max(x,y);}
@@ -14,7 +13,7 @@ public:
         vp c(n);
         for (int i=0; i<n; i++) c[i] = {b[i],a[i]};
         sort(c.rbegin(), c.rend());
-        min_pq<int> pq;
+        min_pq<ll> pq;
         ll ans = 0;
         ll s = 0;
         ll min_b = 1e9;
